Inherit RefBase virtually across libaudioclient
Since RefBase holds a reference count, it should be inherited
virtually in diamond inheritance scenarios. Since the inheritance
structure of callbacks and interfaces is opaque, these classes
should take the slight performance hit of virtual inheritance to
prevent future difficult to detect bugs and leaks. If performance is
critical, we should inherit non-virtually but make the class final.

Bug: 207034551
Test: Compiles
Change-Id: I1ee39c98735757173c81ccdf3dcc09252a073f8a