Update AudioPort <-> configs mapping
It is considered that in general there is a 1:N relationship
between an AudioPort and AudioPortConfigs.
The 'AudioPort.activeConfig' field could only support a 1:1
relationship, but this is true for device ports.
Thus, 'AudioPort.activeConfig' has been moved to 'AudioPortSys'.

The new field 'AudioPortConfig.portId' in the types shared with
HAL is initialized using the ID of the port this config applies
to.

Bug: 205884982
Test: atest audiofoundation_parcelable_test
Change-Id: I8db679213ad582ce106753d254f82ec0318b35bb