Do not send command when aaudio command thread is not running.
If there is error when opening aaudio MMAP stream, the aaudio command
thread may not be started. In that case, if there is any command sending
to the command queue, it will always be blocked until timeout.

In this CL, sending command when the command thread is not running will
directly return AAUDIO_ERROR_INVALID_STATE.

Bug: 209520562
Test: atest AAudioTests
Test: repo steps from the bug
Change-Id: I01c076c120be9a68b984d0c365a61e77821ef117