Use device for given attributes to query direct playback support.
When querying the direct playback support with given audio attributes,
no device information will be taken into consideration when getting
compatible profiles. This will cause the problem that the direct
playback is supported with given audio attributes but the app may not
get direct playback with the same audio attibutes. The reason is that
the routed device(s) may not support direct playback with the audio
attibutes while some other devices do but they are not the routed
devices. In that case, when querying direct playback support, the
devices for given attributes should be used.

Bug: 196047314
Test: atest AudioManagerTest AudioTrackOffloadTest
Change-Id: Idf7d1f5fff44ee542254e72c8251c34a8fe652ff