Improve suppression behavior after reset
Previously, if we had an outlier pose in the buffer right after a
reset, we would consider the state to be "in motion" for the following
window-duration and then on top of that at a window-duration of
suppression. This causes the effective window duration to be twice
the desired one.

With this change, we consider the suppression period to start from the
time of the last outlier pose rather than the time when motion was
last detected.

Test: atest --host libheadtracking-test
Test: Manual verification of head-tracking behavior.

Change-Id: I6b74cc6e7870aa43ca5d79db241b85e0c0217938