Update audiopolicy service to use new sensor privacy api
The new calls no longer require user ids since that is managed
internally in the sensor privacy service and calling services will
receive updates when the state is changed or user is changed who has a
different setting. Calling services should treat the state as a global
setting.

AudioPolicyService doesn't read the state of the microphone toggle
directly so this change shouldn't have any affect.

Test: Build
Bug: 191745272
Change-Id: Ied82f1e91da76706c6522424b7ff273c9a23754f