Unrestrict multilib for extractors.
Libraries should be built as flexibly as possible. Soong won't build
them in any configuration that isn't used. `first` will prevent
multiple 64-bit ABIs from being built in a multi-ABI build, so any
extractor modules cannot be consumed by anything that is
`compile_multilib: "64"` in that configuration if this is `first`.

Test: build/soong/scripts/build-ndk-prebuilts.sh
Bug: http://b/214292894
Change-Id: I7966d419b0d55af1451ae5496f3030bd974b8541