audioflinger: do not spatialize mono or stereo tracks
Never spatialize a mono or stereo track attached to a spatializer
thread. Mono or stereo tracks are never attached to a spatializer thread
unless there is no other choice to reach the selected audio sink device.
In this case it makes more sense to not spatialize the mono or stereo content.

Test: make
Change-Id: Iaf7fd279900bd1873ad8ea4b27e13802e640aaa7