aaudio: clean up MMAP volume TODO
onVolumeChanged() passes in a list of channels and an associated
volume for each one.

However, both the sender (AudioFlinger) and the consumer (AAudio)
is not using this efficiently. AudioFlinger simply sets all the
volumes as the same value and AAudio simply uses the first value.

This CL removes the TODO and cleans up the function parameters.

Bug: 220162278
Test: OboeTester
Change-Id: I3d8b1ad7228e02e2db6d776af7f5a618fdf50a53