vpx: Allow odd width/height decoding
Allow the vpx (vp8 and vp9) decoder components to decode videos
with odd dimensions. This supported by the underlying software
decoders.

We create a graphic buffer that is width aligned to 16 and
height aligned to 2 and then set the "crop" value to the
appropriate video dimensions.

This CL contains changes to files in mainline.

Bug: 224983103
Test: CtsMediaTestCases still passes.
Change-Id: I66fed7df93bc1f5be001209d51a32f9878f78ce7