vpx: Allow odd width/height decoding
Allow the vpx (vp8 and vp9) decoder components to decode videos
with odd dimensions. This supported by the underlying software
decoders.

We create a graphic buffer that is width aligned to 16 and
height aligned to 2 and then set the "crop" value to the
appropriate video dimensions.

This CL contains changes to files not in mainline.

Bug: 224983103
Test: CtsMediaTestCases still passes.
Change-Id: Ia3a3287cf469654297573f0fe1a866ddd8cd27fc