audio policy: fix offload when spatial audio is enabled.
The fact that the spatializer effect is enabled on the output stage
session of the spatializer mixer should not prevent compressed audio
playback from being offloaded.
The logic verifying that non offloadable effects are enabled is
updated to ignore output stage (post processing) effects because those
are not "music effects" but instead apply to all use cases on a given mixer.

Bug: 225119469
Test: enable spatial audio and verify MP3 offload
Change-Id: If612ffb8f7687c345fb0c4671a87f9ab4ad11ea7