[BUG] Fails to find a profile for AudioSource for Direct Output
When trying to start an audio source expected to be routed on a Direct
Output, it fails on identifying the matching profile due to wrong
channel mask.
The channel mask is taken from the source port configuration, which
is an input device. Thus, it fails to apply to identify a sink profile.

Bug: 238058094
Test: adb shell ./data/nativetest/AudioPolicyEmulatorTests/AudioPolicyEmulatorTests --gtest_filter=*AudioSourceBridgingTest.UsingAudioSourceAPI/4

Signed-off-by: François Gaffie <francois.gaffie@renault.com>
Change-Id: I6c460ea765b5eb8ec26e56547e8262beb347f2b2