audio policy: workaround for call signalling tones with hearing aids
When an hearing aid is connected, Telecom will specify stream
BLUETOOTH_SCO whe playing tones with the ToneGenerator.
This causes a mismatch in AudioPolicyManager::checkAndSetVolume()
resulting in muted tones.
Telecom should never use stream BLUETOOTH SCO when creating a
ToneGenerator: this stream type is reserved for volume control.

This CL is a workaround to apply a correct volume until Telecom
fixes this behavior.

Bug: 192520750
Test: repro steps in the bug.
Change-Id: Ie799526aad176563e07faeded83f0495e3b758f5