Increase freshness timeout
Since the head tracker HID spec allows sampling intervals as long as
20ms, setting the freshness timeout to 40ms may result in a staleness
condition if losing one sample.

This change increases to 50ms and also expresses it in terms of time
instead of number of samples, as the actual sample rate might deviate
from the requested one and it becomes confusing.

Bug: This came from a partner request.
Test: Manual verification of SA operation.
Change-Id: I2ec8ceb7cadb4cd8dd4daa681d8770fefdb15f30