audio policy: create spatializer mixer only when needed
Only open a spatializer output when spatial audio is enabled
in order to save memory resources consumed by the dedicated mixer and
audio effect when instantiated.

For the special case where the spatializer output is the only route
available to a specific device, leave it open. This can happen when the
software necoder path is used for Bluetooth A2DP.

Bug: 227740874
Test: make and music playback with and without spatial audio

Change-Id: Id149d0ce000c939c7f45e63b47f2fd1d7a98995a