Remove handle clear on binder death
There is no compelling reason to clear the IEffect handle upon binder
death. Our method calls return fails regardless, and we do not null
check the handle. The notification is also synchronous with existing
calls, leading to a racy use-after-free (sp is not thread-safe). Rather
than incurring the cost of locking, holding onto the handle is fine, as
it is not expensive, and the object will likely be destroyed soon.

Test: Compiles
Bug: 223537153
Change-Id: Id5577d50f0e91c946ef37fde33e3834fec0ca676