camera: make waiting for state more accurate
The current implementation of Camera3Device doesn't account for spurious
thread wakeups when waiting for a specific device state in
`waitUntilStateThenRelock`. This is usually not an issue but in the
worst case can double or triple the time a thread waits for the device
to reach the desired state.

This CL reduces the timeout every time the thread wakes up to ensure
that a thread waits only as long as specified by the passed parameter.

Bug: 223736973
Test: Existing camera CTS tests pass
Change-Id: Ie25c9c8e48f7cd369a089803774d62f6ba566c88