CameraService: Reduce log spam from onCameraAccessPrioritiesChanged.
Bug 231550909 shows an instance of log spam where EX_TRANSACTION_FAILED
is returned from binder. Only one instance of the log is helpful, and
EX_TRANSACTION_FAILED is likely to indicate further failures in the
future, so mute these errors when that happens. Also print the uid and
pid of the failing process to disambiguate the failure, and add
additional checks for the binder status where there previously were
none, so that we may pinpoint the first failed callback.

Bug: 231550909
Test: Smoke tested GCA.
Change-Id: Ibdf54b0d75c0cbc6d147854b0671649bf76a8890