Handle multi-user when disabling MediaMetrics service.
This logic was added in I740b4f5acb88821197e3c4023a4032babd48ec5a but
didn't work for multi-user. Some tests in CtsVoiceInteractionTestCases
where failing on Automotive (in a multi-user environment) because of
additional delays in the test trying to access the MediaMetrics service.

Real world impact: For any secondary user (NOT user 0), the first time
the hotword is spoken after the HotwordDetectionService process starts,
detection fails because the delay causes a timeout in detection. This is
not just a one-time occurrence, as the process may occasionally get
restarted.

Bug: 214488187
Test: `atest CtsVoiceInteractionTestCases` on a raven_car build
Change-Id: If5c721c9acfc78f1cbc35c3056adad95a4604f8b