Camera: Fix rounding issue with minInterval
Existing formula to calculate minInterval may set the minInterval too
large.

For example, if the vSyncInterval is 16.5ms, minFrameDurtion is 33.3ms,
the existing calculation may derive minInterval to be 2+ vSync intervals,
Due to rounding up in calculation, the minInterval could cause the
expectedPresentationTime to be 3 vSyncs. That spaces frames too far away.

Test: Switch between front/back and 30fps/60fps video mode in GCA
Bug: 232722528
Change-Id: I430d571b44a87d32153c54cc661dbf5e42e6ed56