Remove Hermitian tone-mapping curve for HLG
Applying the HLG OOTF as recommended in BT2100 is good enough.

Additionally, fix applying the HLG OOTF to correctly scale to display
brightness. Previously it was scaling to 1000 nits, but renormalize to
[0, 1] assuming a display brightness of 500 nits, which incorrectly
increases luminance.

The net result is that the tone-mapping curve for thumbnails more closely matches the tone-mapping curve used by the default libtonemap implementation used by both display and TextureView, which also just applies the HLG OOTF.

Bug: 232200455
Test: Photos transitions
Change-Id: I03ce481c2f5e47538d1a21f3b6864f61c5a9d9b3