Defer peek legacy configuration until after startup
The OMX codec path will silently drop calls to set parameters until the
codec has been started. By moving the configuration of peek legacy mode
out of kWhatConfigure and into kWhatStartCompleted, this allows the
feature to function on the OMX code path and not just the Codec2 code
path.

Test: add logging for when legacy mode is sent to OMX, and use atest
testTunneledVideoPeekOnAvc with the peek enabled line commented out to
verify that legacy mode is correctly configured in the OMX code path

Bug: 220266078
Bug: 235775488
Change-Id: Ibc8df52cb1b04e787df7f3aa22169e55d0589696